MDEP-689 PR feedback, refactor re-used method to base class, fix newlines
