+++ src/test/resources/unit/go-offline-test/exclude-plugin-config.xml
